Split avatar service modules
